# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API server for Gomaluum, an API that serves i-Ma'luum data for ease of developer.",
        "title": "Gomaluum API Server",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Quddus",
            "url": "http://www.swagger.io/support",
            "email": "ceo@nrmnqdds.com"
        },
        "license": {
            "name": "Bantown Public License",
            "url": "https://github.com/nrmnqdds/gomaluum-api/blob/main/LICENSE.md"
        },
        "version": "2.0"
    },
    "paths": {
...
        "/api/auth/login": {
            "post": {
                "description": "Logs in the user. Save the token and use it in the Authorization header for future requests.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "API key for additional security layer",
                        "name": "x-gomaluum-key",
                        "in": "header"
                    },
                    {
                        "description": "Login properties",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth_proto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.ResponseDTO"
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "get": {
                "description": "Logs out the user. Clears the token from IIUM's CAS. PASETO token is still valid.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "API key for additional security layer",
                        "name": "x-gomaluum-key",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.ResponseDTO"
                        }
                    }
                }
            }
        },
...
        "/api/key/generate": {
            "post": {
                "description": "Generate a new API key for additional authentication layer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "key"
                ],
                "summary": "Generate API Key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API key for additional security layer",
                        "name": "x-gomaluum-key",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.APIKeyResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "auth_proto.LoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "dtos.ResponseDTO": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "errors.CustomError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "server.APIKeyResponse": {
            "type": "object",
            "properties": {
                "api_key": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}